Better move of the test file
