+++ opencti-platform/opencti-graphql/tests/02-integration/05-parser/importCsv-connector/csv-file-cities.csv
