+++ opencti-platform/opencti-graphql/tests/02-integration/05-parser/importCsv-connector/csv-mapper-cities.ts
@@ -1,6 +1,6 @@
-import { type CsvMapperParsed, CsvMapperRepresentationType } from '../../../../../src/modules/internal/csvMapper/csvMapper-types';
-import { ENTITY_TYPE_LOCATION_CITY } from '../../../../../src/schema/stixDomainObject';
-import { ENTITY_TYPE_LABEL } from '../../../../../src/schema/stixMetaObject';
+import { type CsvMapperParsed, CsvMapperRepresentationType } from '../../../../src/modules/internal/csvMapper/csvMapper-types';
+import { ENTITY_TYPE_LOCATION_CITY } from '../../../../src/schema/stixDomainObject';
+import { ENTITY_TYPE_LABEL } from '../../../../src/schema/stixMetaObject';
 
 export const csvMapperMockSimpleCities: Partial<CsvMapperParsed> = {
   id: 'mapper-mock-simple-cities',
