+++ opencti-platform/opencti-graphql/tests/02-integration/05-parser/internal-connectors.json
